+++ packages/taro-components/src/components/swiper/swiper.js
@@ -243,7 +243,16 @@ class Swiper extends Nerv.Component {
         }, this.props.duration)
     )
 
-    if (this.props.onChange) this.props.onChange(ogIndex, currentIndex)
+    if (this.props.onChange) {
+      Object.defineProperty(e, 'detail', {
+        enumerable: true,
+        value: {
+          current: currentIndex,
+          source: 'touch'
+        }
+      })
+      this.props.onChange(e)
+    }
     if (this.props.autoplay) this.pauseAutoPlay()
   }
 
@@ -333,6 +342,17 @@ class Swiper extends Nerv.Component {
         }, this.props.duration)
       }
     )
+    if (this.props.onChange) {
+      let e = new TouchEvent('touchend')
+      Object.defineProperty(e, 'detail', {
+        enumerable: true,
+        value: {
+          current: cur,
+          source: 'autoplay'
+        }
+      })
+      this.props.onChange(e)
+    }
   }
 
   isChangeSlide (translate, currentIndex) {
@@ -420,7 +440,7 @@ class Swiper extends Nerv.Component {
       width: this.state.wrapperWidth,
       height: this.state.wrapperHeight,
       transition: this.state.animating
-        ? `transform ${duration}ms ease-in-out;`
+        ? `transform ${duration}ms ease-in-out`
         : 'none',
       transform: `translate(${!vertical ? this.state.translate : 0}px, ${
         vertical ? this.state.translate : 0