+++ packages/taro/types/index.d.ts
@@ -3,6 +3,23 @@ export as namespace Taro;
 
 declare namespace Taro {
 
+  interface PageNotFoundObject {
+    /**
+     * 不存在页面的路径
+     */
+    path: string,
+
+    /**
+     * 打开不存在页面的 query
+     */
+    query: object,
+
+    /**
+     * 是否本次启动的首个页面（例如从分享等入口进来，首个页面是开发者配置的分享页面）
+     */
+    isEntryPage: boolean
+  }
+
   // Components
   interface ComponentLifecycle<P, S> {
     componentWillMount?(): void;
@@ -14,6 +31,8 @@ declare namespace Taro {
     componentWillUnmount?(): void;
     componentDidShow?(): void;
     componentDidHide?(): void;
+    componentDidCatchError?(err: string): void;
+    componentDidNotFound?(obj: PageNotFoundObject): void;
   }
 
   interface Component<P = {}, S = {}> extends ComponentLifecycle<P, S> {
@@ -21,7 +40,7 @@ declare namespace Taro {
   }
 
   interface PageConfig {
-        /**
+    /**
      * 导航栏背景颜色，HexColor
      * default: #000000
      */
@@ -118,10 +137,16 @@ declare namespace Taro {
     window?: PageConfig
   }
 
+  interface ComponentOptions {
+    addGlobalClass?: boolean
+  }
+
   class Component<P, S> {
     constructor(props?: P, context?: any);
 
     config?: Config;
+
+    options?: ComponentOptions;
 
     $router: {
       params: any
@@ -136,7 +161,7 @@ declare namespace Taro {
 
     render(): any;
 
-    props: Readonly<P>;
+    props: Readonly<P> & Readonly<{ children?: any }>;
     state: Readonly<S>;
     context: any;
     refs: {
@@ -260,7 +285,31 @@ declare namespace Taro {
        *
        * @default false
        */
-      jsonp?: boolean
+      jsonp?: boolean,          
+      /**
+       * 设置H5端 jsonp 请求 url 是否需要被缓存
+       *
+       * @default false
+       */
+      jsonpCache?: boolean,
+      /**
+       * 设置H5端是否允许跨域请求。有效值：no-cors, cors, same-origin
+       *
+       * @default same-origin
+       */
+      mode?: 'no-cors' | 'cors' | 'same-origin',
+      /**
+       * 设置H5端是否携带 Cookie。有效值：include, same-origin, omit
+       *
+       * @default omit
+       */
+      credentials?: 'include' | 'same-origin' | 'omit',
+      /**
+       * 设置H5端缓存模式。有效值：default, no-cache, reload, force-cache, only-if-cached
+       *
+       * @default default
+       */
+      cache?: 'default' | 'no-cache' | 'reload' | 'force-cache' | 'only-if-cached'
     }
   }
   /**
@@ -6597,7 +6646,7 @@ declare namespace Taro {
    *     ```
    * @see https://developers.weixin.qq.com/miniprogram/dev/api/ui-navigate.html#wxnavigatebackobject
    */
-  function navigateBack(OBJECT?: navigateBack.Param): void
+  function navigateBack(OBJECT?: navigateBack.Param): Promise<any>
 
   namespace createAnimation {
     type Param = {
@@ -7230,32 +7279,32 @@ declare namespace Taro {
     fields: (fields: fieldsObject, callback?: fieldCallback) => nodesRef;
     exec: (callback?: execCallback) => void;
   }
-  
+
   interface baseElement {
     id: string,
     dataset: object,
   }
-  
+
   interface rectElement {
     left: number,
     right: number,
     top: number,
     bottom: number,
   }
-  
+
   interface sizeElement {
     width: number,
     height: number,
   }
-  
+
   interface scrollElement {
     scrollLeft: number,
     scrollTop: number
   }
   interface clientRectElement extends baseElement, rectElement, sizeElement {}
-  
+
   interface scrollOffsetElement extends baseElement, scrollElement {}
-  
+
   interface fieldsObject {
     id?:boolean,
     dataset?:boolean,
@@ -7265,18 +7314,18 @@ declare namespace Taro {
     properties?: string[],
     computedStyle?:string[],
   }
-  
+
   interface fieldElement extends baseElement, rectElement, sizeElement {
     [key:string]: any
   }
-  
-  
+
+
   type execObject = clientRectElement & scrollOffsetElement & fieldElement
   type clientRectCallback = (rect: clientRectElement | clientRectElement[]) => void
   type scrollCallback = (res: scrollOffsetElement | scrollOffsetElement[]) => void
   type fieldCallback = (res: fieldElement | fieldElement[]) => void
   type execCallback = (res: execObject | execObject[]) => void
-  
+
   function createSelectorQuery(): SelectorQuery
 
   class SelectorQuery {
@@ -7616,6 +7665,45 @@ declare namespace Taro {
    * @see https://developers.weixin.qq.com/miniprogram/dev/api/open.html#wxgetuserinfoobject
    */
   function getUserInfo(OBJECT?: getUserInfo.Param): Promise<getUserInfo.Promised>
+
+  namespace checkIsSupportFacialRecognition {
+    type Promised = {
+      errMsg: string
+      errCode: number
+    }
+    type Param = {
+      checkAliveType?: number
+    }
+  }
+  function checkIsSupportFacialRecognition(OBJECT?: checkIsSupportFacialRecognition.Param): Promise<checkIsSupportFacialRecognition.Promised>
+
+  namespace startFacialRecognitionVerify {
+    type Promised = {
+      errMsg: string
+      errCode: number
+      verifyResult: string
+    }
+    type Param = {
+      name: string
+      idCardNumber: string
+      checkAliveType?: number
+    }
+  }
+  function startFacialRecognitionVerify(OBJECT?: startFacialRecognitionVerify.Param): Promise<startFacialRecognitionVerify.Promised>
+
+  namespace startFacialRecognitionVerifyAndUploadVideo {
+    type Promised = {
+      errMsg: string
+      errCode: number
+      verifyResult: string
+    }
+    type Param = {
+      name: string
+      idCardNumber: string
+      checkAliveType?: number
+    }
+  }
+  function startFacialRecognitionVerifyAndUploadVideo(OBJECT?: startFacialRecognitionVerifyAndUploadVideo.Param): Promise<startFacialRecognitionVerifyAndUploadVideo.Promised>
 
   namespace requestPayment {
     type Param = {